Rename nrv.consistency.persistence to nrv.consistency.log
I am about to add a new type of consistency persistence not related to transaction log and keeping the old consistency.persistence package name for transaction log is confusing and misleading.
