+++ nrv-core/src/main/scala/com/wajam/nrv/consistency/log/CompositeTransactionLogIterator.scala
@@ -1,4 +1,4 @@
-package com.wajam.nrv.consistency.persistence
+package com.wajam.nrv.consistency.log
 
 import annotation.tailrec
