+++ nrv-core/src/main/scala/com/wajam/nrv/consistency/log/LogRecord.scala
@@ -1,8 +1,8 @@
-package com.wajam.nrv.consistency.persistence
+package com.wajam.nrv.consistency.log
 
 import com.wajam.nrv.utils.timestamp.Timestamp
 import com.wajam.nrv.data.{MessageType, Message}
-import com.wajam.nrv.consistency.persistence.LogRecord.Response.Status
+import com.wajam.nrv.consistency.log.LogRecord.Response.Status
 import scala.language.implicitConversions
 
 sealed trait LogRecord {
