+++ nrv-core/src/main/scala/com/wajam/nrv/consistency/log/NullTransactionLog.scala
@@ -1,7 +1,7 @@
-package com.wajam.nrv.consistency.persistence
+package com.wajam.nrv.consistency.log
 
 import com.wajam.nrv.utils.timestamp.Timestamp
-import com.wajam.nrv.consistency.persistence.LogRecord.Index
+import com.wajam.nrv.consistency.log.LogRecord.Index
 
 object NullTransactionLog extends TransactionLog {
