+++ nrv-core/src/main/scala/com/wajam/nrv/consistency/replication/MasterReplicationSessionManager.scala
@@ -4,7 +4,7 @@ import com.wajam.nrv.service._
 import com.wajam.nrv.utils.timestamp.Timestamp
 import com.wajam.nrv.data._
 import com.wajam.nrv.consistency.{ConsistentStore, ResolvedServiceMember}
-import com.wajam.nrv.consistency.persistence.TransactionLog
+import com.wajam.nrv.consistency.log.TransactionLog
 import com.wajam.nrv.cluster.Node
 import com.wajam.nrv.utils.{CurrentTime, Scheduler, UuidStringGenerator}
 import scala.actors.Actor
