+++ nrv-core/src/test/scala/com/wajam/nrv/consistency/TestTransactionRecorder.scala
@@ -13,8 +13,8 @@ import org.scalatest.matchers.ShouldMatchers._
 
 import com.wajam.nrv.utils.IdGenerator
 import org.scalatest.mock.MockitoSugar
-import persistence.LogRecord.Index
-import com.wajam.nrv.consistency.persistence.{TransactionLogProxy, LogRecord}
+import com.wajam.nrv.consistency.log.LogRecord.Index
+import com.wajam.nrv.consistency.log.{TransactionLogProxy, LogRecord}
 import com.wajam.nrv.utils.timestamp.Timestamp
 
 @RunWith(classOf[JUnitRunner])
