+++ nrv-core/src/test/scala/com/wajam/nrv/consistency/log/TestCompositeTransactionLogIterator.scala
@@ -1,8 +1,8 @@
-package com.wajam.nrv.consistency.persistence
+package com.wajam.nrv.consistency.log
 
 import org.scalatest.matchers.ShouldMatchers._
 import org.mockito.Mockito._
-import com.wajam.nrv.consistency.persistence.LogRecord.Index
+import com.wajam.nrv.consistency.log.LogRecord.Index
 import com.wajam.nrv.consistency.TestTransactionBase
 import org.scalatest.mock.MockitoSugar
