+++ nrv-core/src/test/scala/com/wajam/nrv/consistency/log/TransactionLogProxy.scala
@@ -1,8 +1,8 @@
-package com.wajam.nrv.consistency.persistence
+package com.wajam.nrv.consistency.log
 
 import org.scalatest.mock.MockitoSugar
 import org.mockito.Mockito._
-import com.wajam.nrv.consistency.persistence.LogRecord.Index
+import com.wajam.nrv.consistency.log.LogRecord.Index
 import com.wajam.nrv.utils.timestamp.Timestamp
 import org.mockito.Mockito
